+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - \#2662 - Change Header When Applying Filter in App Collection View
 - \#2663 - Adapt filter count, reflect current result set
 - \#2702 - Show proper info on app creation auth error
+- \#2684 - Running instances are confusing
 
 ## 0.13.5 - 2015-11-24
 ### Fixed